requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.54
botocore==1.35.54
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==5.1.2
djangorestframework==3.15.2
Faker==30.8.1
fonttools==4.54.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pdfcrowd==6.2.1
pdfkit==1.0.0
pillow==10.4.0
psutil==6.1.0
psycopg2-binary==2.9.10
pyarrow==18.0.0
pycparser==2.22
pydyf==0.11.0
pyhtml2pdf==0.0.7
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
s3transfer==0.10.3
selenium==4.26.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
weasyprint==63.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
zopfli==0.2.3.post1
